# docs/api.yaml
openapi: 3.0.3
info:
  title: FinBank — Mini Core Banking System API
  version: 1.0.0
  description: |
    REST API for the Mini Core Banking System.
    Built with FastAPI + MongoDB. Uses JWT Bearer authentication.
  contact:
    name: Team 2
    email: team2@fintech.edu.tr
servers:
  - url: http://localhost:8000
    description: Local Development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CustomerCreate:
      type: object
      required: [first_name, last_name, national_id, phone]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        national_id:
          type: string
          pattern: '^\d{11}$'
        phone:
          type: string
        address:
          type: string
...
    DepositRequest:
      type: object
      required: [account_id, amount]
      properties:
        account_id:
          type: string
        amount:
          type: number
          minimum: 0.01
          maximum: 1000000
        description:
          type: string
    WithdrawRequest:
      type: object
      required: [account_id, amount]
      properties:
        account_id:
          type: string
        amount:
          type: number
          minimum: 0.01
          maximum: 1000000
        description:
          type: string
    TransferRequest:
      type: object
      required: [from_account_id, to_account_id, amount]
      properties:
        from_account_id:
          type: string
        to_account_id:
          type: string
        amount:
          type: number
          minimum: 0.01
          maximum: 1000000
        description:
          type: string
...
paths:
...
  /api/v1/customers/:
    post:
      tags: [Customers]
      summary: Create customer profile (KYC)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerCreate"
      responses:
        "201":
          description: Customer created
    get:
      tags: [Customers]
      summary: List all customers (Admin only)
      responses:
        "200":
          description: List of customers
  /api/v1/customers/me:
    get:
      tags: [Customers]
      summary: Get my customer profile
      responses:
        "200":
          description: Customer profile
        "404":
          description: No profile found
  /api/v1/customers/{customer_id}/status:
    patch:
      tags: [Customers]
      summary: Update KYC status (Admin only)
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status updated
...
  /api/v1/transactions/deposit:
    post:
      tags: [Transactions]
      summary: Deposit money
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Deposit successful
  /api/v1/transactions/withdraw:
    post:
      tags: [Transactions]
      summary: Withdraw money
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequest"
      responses:
        "200":
          description: Withdrawal successful
  /api/v1/transactions/transfer:
    post:
      tags: [Transactions]
      summary: Internal transfer between accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "200":
          description: Transfer completed
...